refactor(backend): colocate rate_limit_exceeded_handler with limiter
Move the 429 JSON handler from main.py into middleware/rate_limit.py
and import it in main alongside the slowapi Limiter.

Made-with: Cursor
