+++ backend/main.py
@@ -1,16 +1,15 @@
 import time
 from contextlib import asynccontextmanager
 
-from fastapi import FastAPI, Request
+from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
-from fastapi.responses import JSONResponse
 from slowapi.errors import RateLimitExceeded
 from slowapi.middleware import SlowAPIMiddleware
 
 import db
 from core.config import STALE_INGESTION_STATUSES, config
 from logging_config.logging_config import setup_logging
-from middleware.rate_limit import limiter
+from middleware.rate_limit import limiter, rate_limit_exceeded_handler
 from middleware.request_id import register_request_id_middleware
 from routes.chat import router as chat_router
 from routes.documents import router as documents_router
@@ -49,22 +48,6 @@ async def lifespan(app: FastAPI):
 
 
 app = FastAPI(lifespan=lifespan)
-
-
-async def rate_limit_exceeded_handler(
-    _request: Request, _exc: RateLimitExceeded
-) -> JSONResponse:
-    return JSONResponse(
-        status_code=429,
-        content={
-            "detail": {
-                "code": "rate_limited",
-                "message": "Too many requests. Please slow down.",
-            }
-        },
-        headers={"Retry-After": "60"},
-    )
-
 
 app.state.limiter = limiter
 app.add_exception_handler(RateLimitExceeded, rate_limit_exceeded_handler)
