+++ backend/middleware/rate_limit.py
@@ -1,6 +1,24 @@
 """Rate limiting setup using slowapi."""
 
+from fastapi import Request
+from fastapi.responses import JSONResponse
 from slowapi import Limiter
+from slowapi.errors import RateLimitExceeded
 from slowapi.util import get_remote_address
 
 limiter = Limiter(key_func=get_remote_address)
+
+
+async def rate_limit_exceeded_handler(
+    _request: Request, _exc: RateLimitExceeded
+) -> JSONResponse:
+    return JSONResponse(
+        status_code=429,
+        content={
+            "detail": {
+                "code": "rate_limited",
+                "message": "Too many requests. Please slow down.",
+            }
+        },
+        headers={"Retry-After": "60"},
+    )
